refactor: create-view UI 컴포넌트 구조 분리
