+++ C3-4T2D/View/Create/Local/CreateMemoView.swift
@@ -1,0 +1,43 @@
+//
+//  CreateMemoView.swift
+//  C3-4T2D
+//
+//  Created by Hwnag Seyeon on 5/30/25.
+//
+
+import SwiftUI
+
+struct CreateMemoView: View {
+    @Binding var descriptionText: String
+
+    var body: some View {
+        VStack(alignment: .leading) {
+            Text("과정에 관한 설명")
+                .font(.system(size: 17, weight: .bold))
+                .foregroundColor(.black)
+                .padding(.bottom, 8)
+
+            ZStack(alignment: .topLeading) {
+                if descriptionText.isEmpty {
+                    Text("과정에 대하여 남기고 싶은 내용이 있으면 적어주세요")
+                        .foregroundColor(.gray)
+                        .padding(.top, 12)
+                        .padding(.leading, 8)
+                        .font(.system(size: 14))
+                }
+
+                TextEditor(text: $descriptionText)
+                    .padding(8)
+                    .font(.system(size: 14))
+                    .scrollContentBackground(.hidden)
+                    .scrollDisabled(true)
+                    .frame(height: 120)
+                    .overlay(
+                        RoundedRectangle(cornerRadius: 6)
+                            .stroke(Color.gray.opacity(0.4), lineWidth: 1)
+                    )
+            }
+        }
+        .padding(.vertical, 16)
+    }
+}
